requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.7.1
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
email-to==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.13.1
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.66.1
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
idna==3.6
lxml==5.3.0
Markdown==3.4.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==10.2.0
platformdirs==4.2.0
premailer==3.10.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.0
python-math==0.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.9.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tesseract==0.1.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.4
urllib3==1.26.16
virtualenv==20.25.0
Wand==0.6.11
webdriver-manager==3.8.6
whitenoise==6.6.0
wsproto==1.2.0
yagmail==0.15.293